Store Remnawave telegram IDs as bigint

+++ panel/backend/app/models.py
@@ -388,7 +388,7 @@ class RemnawaveUser(Base):
     expire_at: Mapped[datetime | None] = mapped_column(DateTime(timezone=True), nullable=True)
     email: Mapped[str | None] = mapped_column(String, nullable=True)
     tag: Mapped[str | None] = mapped_column(String, nullable=True)
-    telegram_id: Mapped[int | None] = mapped_column(Integer, nullable=True)
+    telegram_id: Mapped[int | None] = mapped_column(BigInteger, nullable=True)
     description: Mapped[str | None] = mapped_column(String, nullable=True)
     traffic_limit_bytes: Mapped[int] = mapped_column(BigInteger, default=0, nullable=False)
     traffic_limit_strategy: Mapped[str] = mapped_column(String, default='NO_RESET', nullable=False)
